ai_video_summarizer/requirements.txt
streamlit
google-generativeai
youtube-transcript-api
python-dotenv
